GameOfLife.Scene.Field (game_of_life v0.3.5)

Ist die eingentliche Ansicht des Zellautomaten.

Besteht aus einem Spielfeld aus Buttons, die jeweils eine Zelle representieren. Durch den Button "next" kann der nächte Zustand des Spielfeldes berechnet werden. Durch "run" wird der der Zelleautomat angegewiesen in Intervallen seinen Zustand neu zu berechnen.

Functions

build_lines_h(graph, y, yline)

Specs

build_lines_h(
  graph :: Scenic.Graph.t(),
  y :: pos_integer(),
  yline :: pos_integer()
) :: Scenic.Graph.t()

Zeichent horizontale Linien

Die Linien erzeugen die passenden Rechtecke die dem State des Zellautomaten passen.

build_lines_v(graph, x, xline)

Specs

build_lines_v(
  graph :: Scenic.Graph.t(),
  x :: pos_integer(),
  xline :: pos_integer()
) :: Scenic.Graph.t()

Zeichent vertikale Linien

Die Linien erzeugen die passenden Rechtecke die dem State des Zellautomaten passen.

build_rect(graph, map)

Specs

build_rect(graph :: Scenic.Graph.t(), map :: map()) :: Scenic.Graph.t()

Baut das Zellgitter mit den jeweiligen state des Zellautomaten auf.

Ruft Funktionen für das erstellen des Gitters und zeichnen der aktiven Zellen auf.

fill_rect(graph, xline, yline, list)

Specs

fill_rect(
  graph :: Scenic.Graph.t(),
  xline :: pos_integer(),
  yline :: pos_integer(),
  list()
) :: Scenic.Graph.t()

Zeichnet die aktiven Zellen.

Die Zellen in der Liste werden schwarz gezeichnet.

filter_event(event, from, scene_state)

Specs

filter_event({:click, :next_step}, from :: pid(), state :: term()) ::
  {:noreply, state :: term(), [{:push, g :: Scenic.Graph.t()}]}

filter_event({:click, :intervall}, from :: pid(), state :: term()) ::
  {:noreply, new_state :: term(), [{:push, g :: Scenic.Graph.t()}]}

Verarbeitet die Eingabe eines Buttons.

{:click, :next_step} Lässt den Automaten den neuen Zustand berechnen.

{:click, :intervall} Gibt den Zellautomaten dass Signal automatisch weiter zu laufen oder aufzuhören. Updatete den Button entsprechend.

handle_info(arg1, state)

Specs

handle_info({:new_map, map :: map()}, state :: term()) ::
  {:noreply, new_stat :: term(), [{:push, g :: Scenic.Graph.t()}]}

Updatet den Graph wenn ein neuer Zustand vom Zellautomaten berechnet wurde.

Wartet auf eine Message vom Zellautomaten. Diese enthält einenen aktuellen Graphen. Der aktuelle Graph wird mit build_rect erstellt. Dieser wird anschießend angezeigt. übergibt den State den neuen Zustand des Zellautomaten.

handle_input(event, arg2, scene_state)

Specs

handle_input(
  {:cursor_button,
   {:left, :press, 0, {xposo :: pos_integer(), ypos :: pos_integer()}}},
  context :: Scenic.ViewPort.Context.t(),
  state :: term()
) :: {:noreply, state :: term()}

Ändert den Zustand der Zelle auf die geclickt wurde.

Berrechnet ob der click auf dem Zellgitter getätigt wurde und berechnet die Zelle auf der die Interaktion statt fand.

init(_, opts)

Initialisierung der Oberfläche

Läst das Gitter anhand der Dimensionen aus Agent :xy aufbauen.