Zellautomat (game_of_life v0.3.5)

Logic des Zellautomaten.

Die nötigen Paarameter werden in Agents gespeichert. Der Zelleautomat wird als Genserver Implementiert und von einem Supervisor überwacht. Er kann abstürtzen ohne das ein Datenverlust statt findet.

Functions

alive_in_new_map(k)

Specs

alive_in_new_map(k :: Zelle.t()) :: true | false

Enscheidet über den Zustand der Zelle

Überprüft ob die Zelle den Wert 0 oder 1 bekommt, anhand des eigenen Wertes und der Summe der Nachbarn. Übergibt nur Zellen mit Wert 1 dem :new_map Agent

around_wert(wert, xline, yline, list1, list2, k, torisch)

Specs

around_wert(
  wert :: pos_integer(),
  xline :: pos_integer(),
  yline :: pos_integer(),
  list1 :: list(),
  list2 :: list(),
  k :: Zelle.t(),
  torisch :: boolean()
) :: pos_integer()

Summe der Nachtbarn

Berechnet die Summe der Umligenden Zellen und gibt diese zurück.

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

handle_info(msg, state)

Specs

handle_info({:toggel_cell, z :: [Zelle.t()], pid :: pid()}, state :: term()) ::
  {:noreply, state :: term()}

handle_info({:set_xy, x :: pos_integer(), y :: pos_integer()}, state :: term()) ::
  {:noreply, state :: term()}

handle_info({:new_tick, pid :: pid()}, state :: term()) ::
  {:noreply, state :: term()}

handle_info(
  {:automatic_tick, toggel :: boolean(), pid :: pid()},
  state :: term()
) :: {:noreply, state :: term()}

Alle n Sekunden wird ein neuer Zustand der Zellautomaten berechnet.

Durch den Parameter toggle kann diese Funktion an oder ausgeschaltet werden. Gibt den neuen Zustand zurück.

start_link(opts)

Specs

start_link(_opts :: any()) :: {:ok, pid()}

Initialisierung des Automaten.

Starten des Zellautomaten Als Genserver. Optionale Parameter sind noch nicht vorgesehen.

tick()

Specs

tick() :: :ok

Lässt den nächsten Zustand des Zellautomaten berechnen.

Lässten nächten Zustand berechnen und gibt die neuen Werte an die Agents weiter. Nur Zellen die den Wert 1 haben oder Nachtbar einer Zelle mit dem Wert 1 sind, werden berechnet.

todo_zellen_around(zelle)

Specs

todo_zellen_around(zelle :: Zelle.t()) :: :ok

Ermittelt Nachtbarn und die angegebene Zelle.

Übergibt alle Nachtbarn und die angegebene Zelle, dem :todo Agent. Es sei den, sie befinden sich auserhalb des Zellautomaten.

torus_func(wert, dimension)

Specs

torus_func(wert :: pos_integer(), dimension :: pos_integer()) :: pos_integer()

Setzt eine Zelle auseherhalb der Dimension in den Torischen Bereich