Zustand (game_of_life v0.3.5)

Es Speicher für den Zellautomaten.

Enthält zwei Maps einen für den Aktuellen Zustand und einen für den neuen Zustand des Automaten.

Functions

child_spec(arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

end_tick()

Specs

end_tick() :: :ok

Der neu berechnete Zustand des Zellautomaten wird als aktueller Zustand festgelegt.

Die Map für den Alten Zustand wird gelöscht.

get_akt_cell_wert(z)

Specs

get_akt_cell_wert(z :: Zelle.t()) :: 0 | 1

Gibt den Wert der Aktuellen Zelle zurück.

Wenn die Zelle sich auserhalb der Dimensionen befindet wird Sie auch als 0 betrachtet.

get_akt_map()

Specs

get_akt_map() :: map()

Der aktuelle Zustand wird als Map übergeben.

set_new_cell(z)

Specs

set_new_cell(z :: Zelle.t()) :: :ok

Setzt eine übergebende Zelle mit dem Wert 1 in die Map für den neuen Zustand

start_link(opst)

Specs

start_link(_opts :: any()) :: {:ok, pid()}

Startet die Agents für den Aktuellen und zu berechnenden Zustand.

toggel_cell(z)

Specs

toggel_cell(z :: [Zelle.t()]) :: [true | false]

Verändert die übergebenden Zellen im Aktuellen Zustand des Automaten.

Die übergebenden Zellen werden im Automaten von null auf eins oder umgekert gesetzt.