Zellautomat (game_of_life v0.1.0)

Logic des Zellautomaten. Die nötigen Paarameter werden in Agents gespeichert.

Agents:

xy: hier werden die Dimensionen des Zellautomaten gespeichert es wird auch noch gespeichert ob der Automat von selber laufen soll oder nicht

akt_map: Die Werte der zellen werden als 1 und 0 in einer map mit dem Zellenstruct als id gespeichert um Platz zu sparen werden nach möglichkeit nur zellen mit 1 gespeichert

new_map: dient als Zwischenspicher bei der Berechnung des neuen Zellautomaten

todo: enthält Zellen die zum nächsten Schritt neu berechnet werden müssen

Functions

alive_in_new_map(k)

Specs

alive_in_new_map(k :: Zelle.t()) :: true | false

Enscheidet über den Zustand der Zelle

Überprüft ob die Zelle den Wert 0 oder 1 bekommt, anhand des eigenen Wertes und der Summe der Nachbarn. Übergibt nur Zellen mit Wert 1 dem :new_map Agent

around_wert(wert, xline, yline, list1, list2, k)

Specs

around_wert(
  wert :: pos_integer(),
  xline :: pos_integer(),
  yline :: pos_integer(),
  list1 :: list(),
  list2 :: list(),
  k :: Zelle.t()
) :: pos_integer()

Summe der Nachtbarn

Berechnet die Summe der Umligenden Zellen und gibt diese zurück.

automat()

Specs

automat() :: no_return()

Hauptschleife

Es werde drei Signale verarbeitet die neuen Werte werden über {:new_map, data :: map()} zurück gegeben.

toggel_cell: Es kann geziehlt eine Zelle an oder aus geschaltet werden, abhängig von ihrem aktuellen Zustand.

new_tick: Der nächste Zustand des Automaten wird berechnet.

automatic_tick: Alle n Sekunden wird ein neuer Zustand der Zellautomaten berechnet. Durch den Parameter toggle kann diese Funktion an oder ausgeschaltet werden.

init()

Specs

init() :: no_return()

Initialisierung des Automaten.

Starten der Agenten und setzen der Dimensionen für den Zellautomaten.

tick()

Specs

tick() :: :ok

Lässt den nächsten Zustand des Zellautomaten berechnen.

Lässten nächten Zustand berechnen und gibt die neuen Werte an die Agents weiter. Nur Zellen die den Wert 1 haben oder Nachtbar einer Zelle mit dem Wert 1 sind, werden berechnet.

todo_zellen_around(zelle)

Specs

todo_zellen_around(zelle :: Zelle.t()) :: :ok

Ermittelt Nachtbarn und die angegebene Zelle.

Übergibt alle Nachtbarn und die angegebene Zelle, dem :todo Agent. Es sei den, sie befinden sich auserhalb des Zellautomaten.